+++ src/index.ts
@@ -2,5 +2,6 @@ export {Cache} from './cache';
 export {Cacheable} from './cacheable';
 export {CacheIdFn} from './cache/id/fn';
 export {CacheItem} from './cache/item';
+export {CacheStats} from './cache/stats';
 export {cacheItemId, CacheItemId} from './cache/item/id';
 export {CfgData} from './cfg/data';